Rename directory after compaction is comlete

+++ gobblin-compaction/src/main/java/gobblin/compaction/action/CompactionCompleteFileOperationAction.java
@@ -1,6 +1,9 @@
 package gobblin.compaction.action;
 
+import com.google.common.collect.Lists;
+import gobblin.compaction.dataset.DatasetHelper;
 import gobblin.compaction.mapreduce.CompactionAvroJobConfigurator;
+import gobblin.compaction.mapreduce.MRCompactor;
 import gobblin.compaction.mapreduce.MRCompactorJobRunner;
 import gobblin.compaction.mapreduce.avro.AvroKeyMapper;
 import gobblin.compaction.parser.CompactionPathParser;
@@ -18,6 +21,7 @@ import org.apache.hadoop.mapreduce.Counter;
 import org.apache.hadoop.mapreduce.Job;
 
 import java.io.IOException;
+import java.util.List;
 
 
 /**
@@ -43,36 +47,60 @@ public class CompactionCompleteFileOperationAction implements CompactionComplete
    * Replace the destination folder with new output from map-reduce job
    * and create a file for next run record count comparison.
    */
-  public void onCompactionJobComplete (FileSystemDataset dataset) {
+  public void onCompactionJobComplete (FileSystemDataset dataset) throws IOException {
     if (configurator != null && configurator.isJobCreated()) {
       CompactionPathParser.CompactionParserResult result = new CompactionPathParser(state).parse(dataset);
       Path tmpPath = configurator.getMrOutputPath();
       Path dstPath = new Path (result.getDstAbsoluteDir());
 
-      try {
-        // get record count from map reduce job counter
-        Job job = this.configurator.getConfiguredJob();
-        Counter counter = job.getCounters().findCounter(AvroKeyMapper.EVENT_COUNTER.RECORD_COUNT);
-        long recordCount = counter.getValue();
+      // get record count from map reduce job counter
+      Job job = this.configurator.getConfiguredJob();
+      Counter counter = job.getCounters().findCounter(AvroKeyMapper.EVENT_COUNTER.RECORD_COUNT);
+      long recordCount = counter.getValue();
 
-        // move output from mapreduce to final destination defined by dataset
+
+      boolean renamingRequired = this.state.getPropAsBoolean(MRCompactor.COMPACTION_RENAME_SOURCE_DIR_ENABLED,
+              MRCompactor.DEFAULT_COMPACTION_RENAME_SOURCE_DIR_ENABLED);
+
+      if (renamingRequired) {
+        FsPermission permission = HadoopUtils.deserializeFsPermission(this.state,
+                MRCompactorJobRunner.COMPACTION_JOB_OUTPUT_DIR_PERMISSION,
+                FsPermission.getDefault());
+        WriterUtils.mkdirsWithRecursivePermission(this.fs, dstPath, permission);
+        // append files under mr output to destination
+        List<Path> paths = DatasetHelper.getApplicableFilePaths(fs, tmpPath, Lists.newArrayList("avro"));
+        for (Path path: paths) {
+          String fileName = path.getName();
+          log.info(String.format("Adding %s to %s", path.toString(), dstPath));
+          Path outPath = new Path (dstPath, fileName);
+
+          if (!this.fs.rename(path, outPath)) {
+            throw new IOException(
+                    String.format("Unable to move %s to %s", path.toString(), outPath.toString()));
+          }
+        }
+
+        // update record count (adding previous count)
+        long delta = InputRecordCountHelper.readRecordCount(fs, dstPath);
+        log.info("Will change record count from {} to {} at {}", recordCount, delta, dstPath);
+        recordCount += delta;
+        
+      } else {
         this.fs.delete(dstPath, true);
         FsPermission permission = HadoopUtils.deserializeFsPermission(this.state,
-                                                                      MRCompactorJobRunner.COMPACTION_JOB_OUTPUT_DIR_PERMISSION,
-                                                                      FsPermission.getDefault());
+                MRCompactorJobRunner.COMPACTION_JOB_OUTPUT_DIR_PERMISSION,
+                FsPermission.getDefault());
 
-        WriterUtils.mkdirsWithRecursivePermission (this.fs, dstPath.getParent(), permission);
+        WriterUtils.mkdirsWithRecursivePermission(this.fs, dstPath.getParent(), permission);
         if (!this.fs.rename(tmpPath, dstPath)) {
           throw new IOException(
                   String.format("Unable to move %s to %s", tmpPath, dstPath));
         }
-
-        // write record count
-        InputRecordCountHelper.writeRecordCount (helper.getFs(), new Path (result.getDstAbsoluteDir()), recordCount);
-        log.info("Writing record count {} to {}", recordCount, dstPath);
-      } catch (Exception e) {
-        log.error(e.toString());
       }
+
+      // write record count
+      InputRecordCountHelper.writeRecordCount (helper.getFs(), new Path (result.getDstAbsoluteDir()), recordCount);
+      log.info("Writing record count {} to {}", recordCount, dstPath);
     }
   }
